+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## Unreleased
 
+### Added
+
+- Terminal panes are now focusable. Border glow is shown when focused and cursor changes from an outline to a solid rectangle.
+
 ## [4.3.1] - 2023-10-01
 
 ### Added